requirements.txt
openai==1.12.0
streamlit==1.28.0
chromadb==0.4.22
langchain-text-splitters==0.0.1
PyPDF2==3.0.1
python-dotenv==1.0.0